requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
pandas
plotly
gitpython
numpy
python-multipart
alembic
python-dotenv
aiosqlite
